Remove the logo
I want to add a title parameter so there are no warnings when knitting.
The warnings will be important for the exercise with non-functioning
code. If I add the title, then I will have to add CSS to make sure the
logo displays in the right place. This adds unnecessary complexity to
the example, and it will not work well for the Word output anyways.